+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.0.1 - 2018-10-19
+### Added
+- Added support for advanced language level for Content Quality module
+
+### Changed
+- Page slideout directive has been updated
+
 ## 1.0.0 - 2018-10-11
 ### Added
 - Initial release